cleanups, formatting
All tests pass. One line uncovered, but we
have almost no testing of the JSON functionality,
so this is not unexpected.

DAFFODIL-1298

+++ daffodil-runtime1/src/main/scala/org/apache/daffodil/runtime1/infoset/JsonInfosetOutputter.scala
@@ -139,7 +139,7 @@ class JsonInfosetOutputter private (writer: java.io.BufferedWriter, pretty: Bool
         d.isInfinite || d.isNaN
       }
       case DFDLPrimType.Float => {
-        val f = simple.getFloat.toDouble
+        val f = simple.getFloat
         f.isInfinite || f.isNaN
       }
       case _ => false
